inclusao do teste da api e correcao do metodo update

+++ app/Http/Controllers/TaskController.php
@@ -79,7 +79,10 @@ class TaskController extends Controller
 
         $task->update($request->all());
 
-        $task->tags()->delete();
+        if (count($task->tags()->get()) > 0) {
+            $task->tags()->delete();
+        }
+
         if (count($request->get('tags')) > 0) {
             $_tags = [];
             foreach ($request->get('tags') as $tag) {
